requirements.txt
aerich==0.7.2
aiofiles==23.2.1
aioredis==2.0.1
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
Babel==2.13.1
bcrypt==4.0.1
beautifulsoup4==4.12.3
bleach==6.1.0
click==8.1.7
colorama==0.4.6
dictdiffer==0.9.0
exceptiongroup==1.1.3
fastapi==0.104.1
fastapi-admin==1.0.4
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.1
idna==3.4
iso8601==1.1.0
Jinja2==3.1.2
lxml==5.1.0
MarkupSafe==2.1.3
packaging==23.2
pendulum==2.1.2
py==1.11.0
pydantic==2.4.2
pydantic_core==2.10.1
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
redis==5.0.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
tomlkit==0.12.2
tortoise-orm==0.20.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
webencodings==0.5.1
websockets==12.0
